Last changes may have introduced a bug :
   const char *label; label = 0;

seems to me not the same as :

   QString label; label = QString::null;

because, according to me, the second one makes 'label' an empty string,
and not a string with null address. Therefore the test "if (label)" does
not give the same result as before, if I'm not wrong.

Why do we have to change all "const char *" to QString's ? QString is
slower. Maybe i18n() does now return a QString ?

svn path=/trunk/kdegames/ktuberling/; revision=20973